+++ src/resources/boards/index.ts
@@ -1,5 +1,6 @@
 export { Board } from './board.model';
-export { BoardMemoryRepository as BoardRepository } from './board.memory.repository';
+export { BoardMemoryRepository } from './board.memory.repository';
+export { BoardPostgresRepository as BoardRepository } from './board.postgres.repository';
 export { BoardService } from './board.service';
 export { BoardController } from './board.controller';
 export { createBoardRouter } from './board.router';
