+++ src/resources/tasks/index.ts
@@ -1,5 +1,6 @@
 export { Task } from './task.model';
-export { TaskMemoryRepository as TaskRepository } from './task.memory.repository';
+export { TaskMemoryRepository } from './task.memory.repository';
+export { TaskPostgresRepository as TaskRepository } from './task.postgres.repository';
 export { TaskService } from './task.service';
 export { TaskController } from './task.controller';
 export { createTaskRouter } from './task.router';
