+++ src/resources/users/index.ts
@@ -1,5 +1,6 @@
 export { User } from './user.model';
-export { UserMemoryRepository as UserRepository } from './user.memory.repository';
+export { UserMemoryRepository } from './user.memory.repository';
+export { UserPostgresRepository as UserRepository } from './user.postgres.repository';
 export { UserService } from './user.service';
 export { UserController } from './user.controller';
 export { createUserRouter } from './user.router';
